// frontend/src/models/User/IUserRequest.ts
export interface IUserRequest {
  data: {
    id: number;
    keyData: {
      calorieCount: number;
      carbohydrateCount: number;
      lipidCount: number;
      proteinCount: number;
    };
    todayScore?: number;
    score?: number;
    userInfos: {
      age: number;
      firstName: string;
      lastName: string;
    }
  };
}